Restructure files into 7 levels to provide a better understanding/sense of progression.
