--- src/main/scala/fundamentals/level03/package.scala
@@ -1,15 +1,0 @@
-package fundamentals
-
-package object level03 {
-
-  case class Person(name: String, age: Int)
-
-  sealed trait TrafficLight
-
-  case object Red extends TrafficLight
-
-  case object Yellow extends TrafficLight
-
-  case object Green extends TrafficLight
-
-}
