+++ src/main/scala/introcourse/level02/TypesExercisesNotes.md
