+++ src/main/scala/introcourse/level03/package.scala
@@ -1,0 +1,15 @@
+package introcourse
+
+package object level03 {
+
+  case class Person(name: String, age: Int)
+
+  sealed trait TrafficLight
+
+  case object Red extends TrafficLight
+
+  case object Yellow extends TrafficLight
+
+  case object Green extends TrafficLight
+
+}
