+++ src/main/scala/introcourse/level05/package.scala
@@ -1,0 +1,7 @@
+package introcourse
+
+package object level05 {
+
+  case class Person(name: String, age: Int)
+
+}
